refactor: Broadcast to explorer with failover
